blueprints with context
