+++ app/app.py
@@ -1,11 +1,11 @@
 from flask import Flask
-from app.object.views import object_item
+from app.objects.views import object_item
 from app.pages.views import page
 
 
 # enter point
 def create_app() -> Flask:
-    app = Flask(__name__)
+    app = Flask(__name__, template_folder="../templates")
     register_blueprints(app)
     return app
